# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Reservas - Universidad",
        "description": "Backend para la gestión de reservas de recursos (salas, equipos, etc.) con autenticación JWT y roles admin/usuario.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar todos los usuarios",
                "description": "Obtiene la lista de usuarios (solo administradores). Permite filtrar por estado y rol.",
                "operationId": "a48f811f67f75563295fd522a7d8b8aa",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado (0=inactivo, 1=activo)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "Filtrar por ID del rol",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "juan@universidad.edu"
                                            },
                                            "role_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "estado": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "must_change_password": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "role": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "usuario"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crea un nuevo usuario en el sistema (solo administradores). El usuario creado deberá cambiar su contraseña en el primer ingreso.",
                "operationId": "9feb335faa965a7d5956023a1c641886",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "María García"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@universidad.edu"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6,
                                        "example": "secreto123"
                                    },
                                    "role_id": {
                                        "description": "ID del rol (debe existir en la tabla roles)",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario creado correctamente. Debe cambiar su contraseña al primer ingreso."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "María García"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "maria@universidad.edu"
                                                },
                                                "role_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "estado": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "must_change_password": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/usuarios/perfil": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar perfil del usuario autenticado",
                "description": "Permite al usuario autenticado actualizar su nombre y correo electrónico. No se puede modificar el rol ni el estado.",
                "operationId": "f7b7bc778249042714eb5a1bcde35813",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Mi Nuevo Nombre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mi.nuevo.email@universidad.edu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil actualizado correctamente"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Mi Nuevo Nombre"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "mi.nuevo.email@universidad.edu"
                                                },
                                                "role_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "estado": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación o campos no permitidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}